fix(controller) : add some id route param validator

+++ myride/app/Http/Controllers/Api/InventoryApi/Commands.php
@@ -80,42 +80,52 @@ class Commands extends Controller
     public function hardDeleteInventoryById(Request $request, $id)
     {
         try {
-            $user_id = $request->user()->id;
-
-            // Define user id by role
-            $check_admin = AdminModel::find($user_id);
-            $user_id = $check_admin ? null : $user_id;
-
-            // Get inventory data
-            $inventory = InventoryModel::find($id);
-
-            // Hard Delete inventory by ID
-            $rows = InventoryModel::hardDeleteInventoryById($id, $user_id);
-            if ($rows > 0) {
-                // Delete Firebase uploaded image
-                if ($inventory->inventory_image_url) {
-                    // Delete failed if file not found (already gone)
-                    if (!Firebase::deleteFile($inventory->inventory_image_url)) {
-                        return response()->json([
-                            'status' => 'failed',
-                            'message' => Generator::getMessageTemplate("not_found", 'failed to delete inventory image'),
-                        ], Response::HTTP_NOT_FOUND);
-                    }
-                }
-
-                // Create history
-                HistoryModel::createHistory(['history_type' => 'Inventory', 'history_context' => "remove an inventory"], $user_id);
-
-                // Return success response
-                return response()->json([
-                    'status' => 'success',
-                    'message' => Generator::getMessageTemplate("permentally delete", $this->module),
-                ], Response::HTTP_OK);
-            } else {
+            // Validate param
+            $request->merge(['id' => $id]);
+            $validator = Validation::getValidateId($request);
+            if ($validator->fails()) {
                 return response()->json([
                     'status' => 'failed',
-                    'message' => Generator::getMessageTemplate("not_found", $this->module),
-                ], Response::HTTP_NOT_FOUND);
+                    'message' => $validator->errors()
+                ], Response::HTTP_BAD_REQUEST);
+            } else {
+                $user_id = $request->user()->id;
+
+                // Define user id by role
+                $check_admin = AdminModel::find($user_id);
+                $user_id = $check_admin ? null : $user_id;
+
+                // Get inventory data
+                $inventory = InventoryModel::find($id);
+
+                // Hard Delete inventory by ID
+                $rows = InventoryModel::hardDeleteInventoryById($id, $user_id);
+                if ($rows > 0) {
+                    // Delete Firebase uploaded image
+                    if ($inventory->inventory_image_url) {
+                        // Delete failed if file not found (already gone)
+                        if (!Firebase::deleteFile($inventory->inventory_image_url)) {
+                            return response()->json([
+                                'status' => 'failed',
+                                'message' => Generator::getMessageTemplate("not_found", 'failed to delete inventory image'),
+                            ], Response::HTTP_NOT_FOUND);
+                        }
+                    }
+
+                    // Create history
+                    HistoryModel::createHistory(['history_type' => 'Inventory', 'history_context' => "remove an inventory"], $user_id);
+
+                    // Return success response
+                    return response()->json([
+                        'status' => 'success',
+                        'message' => Generator::getMessageTemplate("permentally delete", $this->module),
+                    ], Response::HTTP_OK);
+                } else {
+                    return response()->json([
+                        'status' => 'failed',
+                        'message' => Generator::getMessageTemplate("not_found", $this->module),
+                    ], Response::HTTP_NOT_FOUND);
+                }
             }
         } catch(\Exception $e) {
             return response()->json([
